+++ data-extraction/eicu_patient_cohort.sql
@@ -12,7 +12,7 @@ SELECT
   pat.uniquepid AS subject_id,
   pat.patienthealthsystemstayid AS hadm_id,
   pat.patientunitstayid AS icustay_id,
-  SAFE_CAST(REGEXP_EXTRACT(pat.age, r"[0-9]+") as float) AS age,
+  SAFE_CAST(REGEXP_EXTRACT(pat.age, r"[0-9]+") as FLOAT64) AS age,
   pat.hospitaladmitoffset AS hospitaladmitoffset,
   pat.unitdischargeoffset / (24 * 60) AS icu_length_of_stay,
   pat.hospitalid AS hospital_id
@@ -37,7 +37,7 @@ WHERE pat.unitdischargeoffset > 0),
 
 ventilation AS (
 SELECT
-  MAX(SAFE_CAST(chart.nursingchartvalue as float)) AS max_fiO2,
+  MAX(SAFE_CAST(chart.nursingchartvalue as FLOAT64)) AS max_fiO2,
   chart.patientunitstayid AS icustay_id
 FROM chart
 WHERE chart.nursingchartcelltypevalname = "O2 L/%"
